Refactor: 진행중 심부름 확인 관련 기능을 StatusMessageController로 이동
