+++ server/appteam/src/main/java/com/pknuErrand/appteam/service/statusMessage/StatusMessageService.java
@@ -5,12 +5,14 @@ import com.pknuErrand.appteam.domain.errand.Errand;
 import com.pknuErrand.appteam.domain.errand.ErrandCompletionStatus;
 import com.pknuErrand.appteam.domain.member.Member;
 import com.pknuErrand.appteam.domain.statusMessage.StatusMessage;
+import com.pknuErrand.appteam.dto.errand.getDto.InProgressErrandListResponseDto;
 import com.pknuErrand.appteam.dto.statusMessage.StatusMessageRequestDto;
 import com.pknuErrand.appteam.dto.statusMessage.StatusMessageResponseDto;
 import com.pknuErrand.appteam.exception.CustomException;
 import com.pknuErrand.appteam.exception.ErrorCode;
 import com.pknuErrand.appteam.repository.StatusMessageRepository;
 import com.pknuErrand.appteam.repository.errand.ErrandCompletionStatusRepository;
+import com.pknuErrand.appteam.repository.errand.ErrandRepository;
 import com.pknuErrand.appteam.service.errand.ErrandService;
 import com.pknuErrand.appteam.service.member.MemberService;
 import org.springframework.stereotype.Service;
@@ -24,16 +26,19 @@ import java.util.List;
 public class StatusMessageService {
     private final StatusMessageRepository statusMessageRepository;
     private final ErrandCompletionStatusRepository errandCompletionStatusRepository;
+    private final ErrandRepository errandRepository;
     private final MemberService memberService;
     private final ErrandService errandService;
     public StatusMessageService(StatusMessageRepository statusMessageRepository,
                                 ErrandCompletionStatusRepository errandCompletionStatusRepository,
+                                ErrandRepository errandRepository,
                                 MemberService memberService,
                                 ErrandService errandService) {
         this.statusMessageRepository = statusMessageRepository;
         this.memberService = memberService;
         this.errandService = errandService;
         this.errandCompletionStatusRepository = errandCompletionStatusRepository;
+        this.errandRepository = errandRepository;
     }
     public StatusMessage saveStatusMessage(StatusMessageRequestDto statusMessageRequestDto, Long errandNo) {
         Member errander = memberService.findMemberByMemberNo(statusMessageRequestDto.getErranderNo());
@@ -59,6 +64,35 @@ public class StatusMessageService {
         }
         return filteredStatusMessageList;
     }
+
+    public void checkInProgressErrandExist() {
+        Member member = memberService.getLoginMember();
+        List<Errand> inProgressErrandList = errandRepository.findInProgressErrand(member.getMemberNo());
+        if(inProgressErrandList.size() == 0)
+            throw new CustomException(ErrorCode.ERRAND_NOT_FOUND);
+//        for(Errand errand: inProgressErrandList)
+//            System.out.println(errand.getErrandNo());
+    }
+
+    @Transactional
+    public List<InProgressErrandListResponseDto> getInProgressErrand() {
+        Member member = memberService.getLoginMember();
+        List<Errand> inProgressErrandList = errandRepository.findInProgressErrand(member.getMemberNo());
+        if(inProgressErrandList.size() == 0)
+            throw new CustomException(ErrorCode.ERRAND_NOT_FOUND);
+        List<InProgressErrandListResponseDto> filteredinProgressErrandList = new ArrayList<>();
+        for(Errand errand : inProgressErrandList) {
+            InProgressErrandListResponseDto filteredErrand = InProgressErrandListResponseDto.builder()
+                    .errandNo(errand.getErrandNo())
+                    .title(errand.getTitle())
+                    .due(errand.getDue())
+                    .isUserOrder(errand.getOrderNo().equals(member))
+                    .build();
+            filteredinProgressErrandList.add(filteredErrand);
+        }
+        return filteredinProgressErrandList;
+    }
+
 
     public void erranderCompletionConfirm(Long errandNo) {
         Member errander = memberService.getLoginMember();
